+++ flutter/lib/src/models.dart
@@ -338,7 +338,7 @@ class BubblFlushResult {
 
 class BubblDiagnostics {
   const BubblDiagnostics({
-    this.sdkVersion = '4.0.0',
+    this.sdkVersion = '4.0.1',
     this.platform = 'flutter',
     this.booted = false,
     this.pendingIngestCount = 0,